pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "moodle-to-vikwikiquiz"
description = "A CLI for converting graded Moodle quiz HTMLs to a vik.wiki quiz wikitext"
readme = "README.md"
authors = [
    { name="Máté Gyöngyösi", email="melees00_quern@icloud.com" },
]
version = "1.2.2"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Typing :: Typed",
    "Framework :: Pydantic",
    "Intended Audience :: Education",
]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4",
    "plum-dispatch",
    "pylatexenc",
    "pyperclip",
]

[project.urls]
Homepage = "https://github.com/gy-mate/moodle-to-vikwikiquiz"
Issues = "https://github.com/gy-mate/moodle-to-vikwikiquiz/issues"
Funding = "https://github.com/sponsors/gy-mate"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
moodle-to-vikwikiquiz = "moodle_to_vikwikiquiz.main:main"
